[Feat] Moved Interpolate to math module, added LinAlg2D
