+++ CMakeLists.txt
@@ -72,6 +72,7 @@ add_subdirectory(src/loki/gui)
 add_subdirectory(src/loki/hash)
 add_subdirectory(src/loki/input)
 add_subdirectory(src/loki/json)
+add_subdirectory(src/loki/math)
 add_subdirectory(src/loki/screens)
 add_subdirectory(src/loki/shapes)
 add_subdirectory(src/loki/sprites)
