+++ src/loki/common/All.hpp
@@ -5,7 +5,6 @@
  */
 #pragma once
 
-#include "Interpolate.hpp"
 #include "IterAdapters.hpp"
 #include "Matrix.hpp"
 #include "Matrix.inl"
