+++ src/loki/math/All.hpp
@@ -1,0 +1,9 @@
+/*!
+ * \file All.hpp
+ * \author Srykah
+ * \copyright GNU GPL v3.0
+ */
+#pragma once
+
+#include "Interpolate.hpp"
+#include "LinAlg2D.hpp"
