+++ src/loki/math/LinAlg2D.cpp
@@ -1,0 +1,12 @@
+/*!
+ * \file LinAlg2D.cpp.c
+ * \author Srykah
+ * \copyright GNU GPL v3.0
+ */
+#include "LinAlg2D.hpp"
+
+namespace loki::math {
+
+
+
+}
